+++ packages/api-reference/src/components/Content/Schema/Schema.vue
@@ -27,20 +27,6 @@ const shouldShowToggle = computed(() => {
   return true
 })
 
-// Merge the (optional) `additionalProperties` into the schema
-const mergedSchema = computed(() => {
-  return {
-    ...(typeof props.value === 'object' ? props.value : {}),
-    ...(typeof props.value === 'object' &&
-    typeof props.value?.additionalProperties === 'object'
-      ? {
-          ...props.value?.additionalProperties,
-          additionalProperties: undefined,
-        }
-      : {}),
-  }
-})
-
 // Prevent click action if noncollapsible
 const handleClick = (e: MouseEvent) =>
   props.noncollapsible && e.stopPropagation()
@@ -99,30 +85,56 @@ const handleClick = (e: MouseEvent) =>
               size="md" />
             <SchemaHeading
               :name="name"
-              :value="mergedSchema" />
+              :value="value" />
           </template>
         </DisclosureButton>
         <DisclosurePanel :static="noncollapsible">
-          <template v-if="mergedSchema?.properties">
-            <SchemaProperty
-              v-for="property in Object.keys(mergedSchema?.properties)"
-              :key="property"
-              :compact="compact"
-              :level="level"
-              :name="property"
-              :required="
-                mergedSchema.required &&
-                mergedSchema.required.length &&
-                mergedSchema.required.includes(property)
-              "
-              :value="mergedSchema.properties?.[property]" />
+          <template v-if="value.properties || value.additionalProperties">
+            <template v-if="value.properties">
+              <SchemaProperty
+                v-for="property in Object.keys(value?.properties)"
+                :key="property"
+                :compact="compact"
+                :level="level"
+                :name="property"
+                :required="
+                  value.required &&
+                  value.required.length &&
+                  value.required.includes(property)
+                "
+                :value="value.properties?.[property]" />
+            </template>
+            <template v-if="value.additionalProperties">
+              <!-- 
+                Allows any type of additional property value
+                @see https://swagger.io/docs/specification/data-models/dictionaries/#free-form
+               -->
+              <SchemaProperty
+                v-if="
+                  value.additionalProperties === true ||
+                  Object.keys(value.additionalProperties).length === 0
+                "
+                additional
+                :compact="compact"
+                :level="level"
+                noncollapsible
+                :value="{ type: 'any' }" />
+              <!-- Allows a specific type of additional property value -->
+              <SchemaProperty
+                v-else
+                additional
+                :compact="compact"
+                :level="level"
+                noncollapsible
+                :value="value.additionalProperties" />
+            </template>
           </template>
           <template v-else>
             <SchemaProperty
               :compact="compact"
               :level="level"
-              :name="mergedSchema?.name"
-              :value="mergedSchema" />
+              :name="value.name"
+              :value="value" />
           </template>
         </DisclosurePanel>
       </div>