+++ packages/api-reference/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scalar/api-reference
 
+## 1.22.5
+
+### Patch Changes
+
+- d03ab0f: feat: add better support for additional properties
+
+## 1.22.4
+
+### Patch Changes
+
+- adc0693: feat: use http as default scheme for Swagger 2.0 specifications
+
 ## 1.22.3
 
 ### Patch Changes