+++ packages/mock-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scalar/mock-server
 
+## 0.1.40
+
+### Patch Changes
+
+- Updated dependencies [d03ab0f]
+  - @scalar/api-reference@1.22.5
+
+## 0.1.39
+
+### Patch Changes
+
+- Updated dependencies [adc0693]
+  - @scalar/api-reference@1.22.4
+
 ## 0.1.38
 
 ### Patch Changes