+++ packages/nextjs-api-reference/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scalar/nextjs-api-reference
 
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies [d03ab0f]
+  - @scalar/api-reference@1.22.5
+
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [adc0693]
+  - @scalar/api-reference@1.22.4
+
 ## 0.3.3
 
 ### Patch Changes