+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scalar/nuxt
 
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [d03ab0f]
+  - @scalar/api-reference@1.22.5
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [adc0693]
+  - @scalar/api-reference@1.22.4
+
 ## 0.0.12
 
 ### Patch Changes